Fix some tests

+++ app/src/test/java/org/hwyl/sexytopo/control/util/Space2DUtilsTest.java
@@ -40,7 +40,7 @@ public class Space2DUtilsTest {
         path.add(new Coord2D(0,0));
         path.add(new Coord2D(10,0));
 
-        List<Coord2D> simplifiedPath = Space2DUtils.simplify(path, 0.01);
+        List<Coord2D> simplifiedPath = Space2DUtils.simplify(path, 0.01f);
 
         Assert.assertEquals(path, simplifiedPath);
     }
@@ -56,7 +56,7 @@ public class Space2DUtilsTest {
         path.add(new Coord2D(40,0));
         path.add(new Coord2D(50,0));
 
-        double epsilon = Space2DUtils.simplificationEpsilon(50, 50);
+        float epsilon = Space2DUtils.simplificationEpsilon(50, 50);
         List<Coord2D> simplifiedPath = Space2DUtils.simplify(path, epsilon);
 
         ArrayList<Coord2D> expectedPath = new ArrayList<>();
@@ -76,7 +76,7 @@ public class Space2DUtilsTest {
         path.add(new Coord2D(10,5));
         path.add(new Coord2D(10,10));
 
-        double epsilon = Space2DUtils.simplificationEpsilon(10, 10);
+        float epsilon = Space2DUtils.simplificationEpsilon(10, 10);
         List<Coord2D> simplifiedPath = Space2DUtils.simplify(path, epsilon);
 
         ArrayList<Coord2D> expectedPath = new ArrayList<>();
@@ -90,23 +90,23 @@ public class Space2DUtilsTest {
     @Test
     public void testSimplifyDensifiedCircle() {
 
-        double pi2 = 2 * Math.PI;
-        double step = pi2/12;
-        double radius = 5;
+        float pi2 = 2 * (float)Math.PI;
+        float step = pi2/12;
+        float radius = 5;
         int offset = 10;
 
         // Make a 12 point circle approximation
         ArrayList<Coord2D> path = new ArrayList<>();
         for(double theta = 0; theta < pi2; theta += step)
         {
-            double x = offset + radius * Math.cos(theta);
-            double y = offset - radius * Math.sin(theta);
+            float x = offset + radius * (float)Math.cos(theta);
+            float y = offset - radius * (float)Math.sin(theta);
             path.add(new Coord2D(x, y));
         }
         path.add(path.get(0));
 
         {
-            double epsilon = Space2DUtils.simplificationEpsilon(10, 10);
+            float epsilon = Space2DUtils.simplificationEpsilon(10, 10);
             List<Coord2D> simplifiedPath = Space2DUtils.simplify(path, epsilon);
 
             // Remains unchanged
@@ -114,7 +114,7 @@ public class Space2DUtilsTest {
         }
 
         {
-            double epsilon = 4;
+            float epsilon = 4;
             List<Coord2D> simplifiedPath = Space2DUtils.simplify(path, epsilon);
 
             // Aggressive epsilon simplifies to a diamond
